learn/requirements.txt
jsonfield==2.0.2
graphene-django==1.3
django-reversion==2.0.10
django-extensions==1.9.1
django-papertrail==1.1.4